+++ src/components/Userpage/AuthPassword.js
@@ -104,4 +104,4 @@ const Container = styled.div`
   }
 `;
 
-export default AuthPassword;
+export { AuthPassword };
