+++ src/components/Userpage/EditProfile.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import styled from 'styled-components';
 
-const EditMemberInfo = () => {
+const EditProfile = () => {
   return <Container>내정보 수정 컴포넌트</Container>;
 };
 
 const Container = styled.div``;
 
-export default EditMemberInfo;
+export { EditProfile };
