+++ src/components/Userpage/OrderHistory.js
@@ -1,10 +1,11 @@
 import React, { useEffect, useState } from 'react';
-import { Link } from 'react-router-dom';
+import { Link, useParams } from 'react-router-dom';
 import reset from '../../css/reset-css.css';
 import styled from 'styled-components';
 import { orderedProducts, confirmProduct, refundProduct } from '../../utils/useAPI';
 
 const OrderHistory = () => {
+  const { menu } = useParams();
   const [ordered, setOdered] = useState([]);
 
   useEffect(() => {
@@ -46,6 +47,7 @@ const OrderHistory = () => {
 
   return (
     <Container>
+      <h2>{menu}페이지 입니다.</h2>
       <ol>
         {ordered.map((list) => (
           <li key={list.detailId} className="orderedList">
@@ -170,4 +172,4 @@ const CancleOk = styled.div`
   }
 `;
 
-export default OrderHistory;
+export { OrderHistory };
