+++ src/pages/User/index.js
@@ -1,0 +1,5 @@
+export * from '../../components/Userpage/OrderHistory';
+export * from '../../components/Userpage/CancelHistory';
+export * from '../../components/Userpage/BankAccounts';
+export * from '../../components/Userpage/EditProfile';
+export * from '../../components/Userpage/AuthPassword';
