+++ src/pages/index.js
@@ -5,4 +5,4 @@ export * from './OrderPage';
 export * from './Auth/LoginPage';
 export * from './Auth/SignUpPage';
 export * from './Category/CategoryPage';
-export * from './UserPage';
+export * from './User/UserPage';
